refactoring into new structure
