+++ brain-defence/arcade/levels/base_map.py
@@ -1,0 +1,7 @@
+
+#
+# class BaseMap():
+#
+#     def __init__(self):
+#         self.active_impressions = []
+#         #impressions_spawn_plan
